feat: refactor code evaluation handling to utilize codeeval package and improve response management
